REFACTOR: moves logic to organize payments into years into helper function, moves jsx portion of helper function into MortgageScheduleYear, updates ScheduledPayment type definition, removing jsx property
